+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/DefaultEntityComponentFactory.java
@@ -28,7 +28,7 @@ import is.codion.swing.framework.model.SwingEntityEditModel;
 import javax.swing.JComponent;
 import java.time.temporal.Temporal;
 
-import static is.codion.swing.common.ui.component.Components.fileInputPanel;
+import static is.codion.swing.common.ui.component.Components.byteArrayInputPanel;
 import static is.codion.swing.framework.ui.component.EntityComponents.entityComponents;
 import static java.util.Objects.requireNonNull;
 
@@ -59,7 +59,7 @@ public class DefaultEntityComponentFactory<T, C extends JComponent> implements E
 			return createTemporalComponentValue(attribute, (Temporal) initialValue, inputComponents);
 		}
 		if (attribute.type().isByteArray()) {
-			return (ComponentValue<T, C>) fileInputPanel()
+			return (ComponentValue<T, C>) byteArrayInputPanel()
 							.buildValue();
 		}